+++ src/__mockData__/swissprot_entry.json
