+++ src/results/__tests__/ResultsContainer.spec.tsx
@@ -5,8 +5,8 @@ import 'core-js/stable';
 import { cleanup, waitForElement, fireEvent } from '@testing-library/react';
 import ResultsContainer from '../ResultsContainer';
 import { act } from 'react-dom/test-utils';
-import results from '../../../__mockData__/results.json';
-import entry from '../../../__mockData__/swissprot_entry.json';
+import results from '../../__mockData__/results.json';
+import entry from '../../__mockData__/swissprot_entry.json';
 import searchInitialState from '../../search/state/searchInitialState';
 import resultsInitialState, { ViewMode } from '../state/resultsInitialState';
 import renderWithRedux from '../../__testHelpers__/renderWithRedux';
